Remove lazyredraw option
In trying to track down a bug causing stray/phantom characters
requiring a redraw, I remember that I set lazyredraw because of
a tip that it makes high count macros faster. But, I rarely use
macros in that way so I'm turning it back off in case it's the
cause.