Fix JS template literal highlighting
Either the JavaScript and/or the JSX syntax plugins loaded with
Polyglot are cause some syntax highlighting problems with template
literals. There are open issues in both repos and I'm unsure of
where the underlying issue is. But disabling JSX via the Polyglot
settings works as a temporary fix for now even though I lose HTML
template highlighting within literals.